+++ pyengine/Utils/__init__.py
@@ -1,4 +1,4 @@
-from pyengine.Utils.Others import clamp, wrap_text
+from pyengine.Utils.Others import clamp, wrap_text, get_images_from_gif
 from pyengine.Utils.Font import Font
 from pyengine.Utils.Color import Color, Colors
 from pyengine.Utils.Vec2 import Vec2